# Update Node from 14 to 18; update other dev package versions