--- a/app/templates/courses/teacher-classes-view.coco.pug
+++ b/app/templates/courses/teacher-classes-view.coco.pug
@@ -36,6 +36,7 @@ block content
     - var officeHours = view.upcomingOfficeHours;
     - var limit = view.howManyOfficeHours == 'all' ? Infinity : 2;
     - var limited = officeHours.length > 2;
+    - var limitedOfficeHours = officeHours.slice(0, limit);
     #office-hours
       if officeHours.length && view.howManyOfficeHours != 'none'
         .update-class.rtl-allowed
@@ -49,7 +50,7 @@ block content
               span(data-i18n="teacher.office_hours_link")
               span.spl(data-i18n="teacher.office_hours_detail_2")
             ul.list-group
-              for oh in officeHours.slice(0, limit)
+              for oh in limitedOfficeHours
                 li.list-group-item
                   span.spr #{moment(oh.time).calendar(null, {sameElse: 'ddd MMM D, LT'})}:
                   a(href=oh.link target="_blank")= oh.name