--- a/package.json
+++ b/package.json
@@ -213,7 +213,7 @@
     "nodemon": "^2.0.19",
     "pg": "^6.1.2",
     "portfinder": "^1.0.28",
-    "pug": "^3.0.2",
+    "pug": "^2.0.3",
     "pug-lint": "^2.6.0",
     "pug-loader": "^2.4.0",
     "recursive-readdir-sync": "^1.0.6",