Between updating Node, updating some of the other dev package versions, and setting dev build source maps to faster, I was able to get full compile times on my machine down from 73s to 38s and rebuild times from 3s to 1.1s.

With generous use of IgnorePlugin for areas of the code I wasn't working on, I got PlayLevelView-related compiles to 25s / 0.6s.

One annoyance: we must use `npm install --legacy-peer-deps` because `pug-loader` hasn't been updated and otherwise throws an error. Maybe we can either switched to a patched version of `pug-loader` or switch to a more up-to-date, full-featured one like `pug-plugin`. Should figure that out before merging this in, and likely do other work to update our node version more generally.